Rename data_structures module to models
